reorganize cli tests
